requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
mysql-connector-python==8.1.0
requests==2.31.0
python-socketio==5.8.0
eventlet==0.33.3
